[vtk] customize offset/connectivity fields
We noticed that (with the dune adapters) significant CPU time is
spent determining the field layouts, because the dune entity ranges
for partitions are not "sized" and we have to iterate multiple times
through the entire grid (in every call to layout() and serialized()).
This is now changed to custom field implementations that compute and
cache the layout upon construction. We should not run into issues for
dynamically changing grids (e.g. over time), since these fields are
created by VTK writers just before the actual writing, and go out of
scope again afterwards.