[vtk][xml] avoid compressor template arg
It is unlikely that the number of compressors supported by VTK is
something that will frequently change. Therefore, we remove the
compressor from the template args and switch to a variant instead.
This also facilitates storing instances of writers as one does not
have to know as many template arguments.