[grid][discontinuous] support rvalue refs
This commit introduces changes in order to make the discontinuous grid
compatible with grids that yield temporaries for cells, in conjunction
with CellPoints trait implementations that use lambdas that capture
references to these cells. These references would be dangling for the
discontinuous grid which internally caches the result of CellPoints.
Therefore, in those cases it also caches the current cell.